+++ src/main/scala/MyAux.scala
@@ -1,0 +1,3 @@
+//case class MyAux[] {
+//
+//}
