+++ src/main/scala/VariableResolutionImplicits.scala
@@ -1,5 +1,7 @@
 package slender
 
+import shapeless._
+
 /**Base trait for resolvers.
   * Instances of ResolverBase may be provided by the implicit CantResolve method, which which will be provide a
   * 'do-nothing' resolver for any expression.
@@ -16,13 +18,18 @@ package slender
   * This has the side-effect of not being able to tell if the result of an outermost For comprehension is actually
   * resolved without either calling 'isResolved' or explicitly re-resolving, but its a good interim solution.
 */
-trait ResolverBase[-In <: Expr,+Out <: Expr] extends (In => Out) with Serializable
+trait ResolverBase[-In,+Out] extends (In => Out) with Serializable
 
-trait Resolver[-In <: Expr,+Out <: Expr] extends ResolverBase[In,Out]
+trait Resolver[-In,+Out] extends ResolverBase[In,Out]
 
 trait Binder[V <: VariableExpr[V],T,-In <: Expr,+Out <: Expr] extends (In => Out) with Serializable
 
 object Resolver {
+
+  def instance[In, Out](f: In => Out): Resolver[In,Out] = new Resolver[In,Out] {
+    def apply(v1: In): Out = f(v1)
+  }
+
   def nonResolver[E <: Expr]: Resolver[E,E] = new Resolver[E,E] { def apply(v1: E) = v1 }
 }
 
@@ -32,7 +39,7 @@ object Binder {
 }
 
 
-trait LowPriorityResolutionImplicits {
+trait Priority0ResolutionImplicits {
   //A default ResolverBase for any expression, so that for-comprehensions in the DSL can resolve expressions before
   //returning them (inner expressions may not be resolvable when initially returned).
   //Note that the other methods here and the main Resolve method require a Resolver not a ResolverBase so none of
@@ -40,7 +47,58 @@ trait LowPriorityResolutionImplicits {
   implicit def CantResolve[E <: Expr]: ResolverBase[E,E] = new ResolverBase[E,E] { def apply(v1: E) = v1 }
 }
 
-trait StandardPriorityResolutionImplicits extends LowPriorityResolutionImplicits {
+trait Priority01ResolutionImplicits extends Priority0ResolutionImplicits {
+
+//  implicit def HNilResolver: Resolver[GenExpr[HNil], GenExpr[HNil]] =
+//    Resolver.instance[GenExpr[HNil], GenExpr[HNil]](identity[GenExpr[HNil]])
+//
+//  implicit def HListResolver[H1 <: Expr, T1 <: HList, H2 <: Expr, T2 <: HList]
+//  (implicit resolveH: Resolver[H1, H2], resolveT: Resolver[GenExpr[T1], GenExpr[T2]]):
+//    Resolver[GenExpr[H1 :: T1], GenExpr[H2 :: T2]] = Resolver.instance[GenExpr[H1 :: T1], GenExpr[H2 :: T2]] {
+//    case GenExpr(h1 :: t1) => {
+//      val h2 = resolveH(h1)
+//      val t2 = resolveT(GenExpr(t1)).repr
+//      val hlist2 = h2 :: t2
+//      GenExpr(hlist2)
+//    }
+//  }
+  implicit def HNilResolver: Resolver[HNil,HNil] =
+    Resolver.instance(identity[HNil])
+
+  implicit def HListResolver[H1 <: Expr, T1 <: HList, H2 <: Expr, T2 <: HList]
+  (implicit resolveH: Resolver[H1, H2], resolveT: Resolver[T1, T2]):
+    Resolver[H1 :: T1, H2 :: T2] = Resolver.instance[H1 :: T1, H2 :: T2] {
+    case h1 :: t1 => resolveH(h1) :: resolveT(t1)
+  }
+
+  implicit def InductiveResolver2[
+    C[_ <: Expr,_ <: Expr],L1 <: Expr,R1 <: Expr,L2 <: Expr,R2 <: Expr
+  ]
+  (implicit gen1: Generic.Aux[C[L1,R1],L1::R1::HNil], genResolve: Resolver[L1::R1::HNil,L2::R2::HNil],
+   gen2: Generic.Aux[C[L2,R2], L2::R2::HNil]):
+  Resolver[C[L1,R1], C[L2,R2]] = new Resolver[C[L1,R1], C[L2,R2]] {
+    def apply(v1: C[L1,R1]): C[L2, R2] = gen2.from(genResolve(gen1.to(v1)))
+  }
+
+  implicit def InductiveResolver1[
+    C[_ <: RingExpr],E1 <: RingExpr,E2 <: RingExpr
+  ]
+  (implicit gen1: Generic.Aux[C[E1],E1::HNil], genResolve: Resolver[E1::HNil,E2::HNil],
+   gen2: Generic.Aux[C[E2], E2::HNil]):
+  Resolver[C[E1], C[E2]] = new Resolver[C[E1], C[E2]] {
+    def apply(v1: C[E1]): C[E2] = gen2.from(genResolve(gen1.to(v1)))
+  }
+
+//  implicit def InductiveResolver2[
+//  C[_,_],L1 <: Expr,R1 <: Expr,L2 <: Expr,R2 <: Expr, Repr1 <: HList, Repr2 <: HList
+//  ]
+//  (implicit gen1: Generic.Aux[C[L1,R1],Repr1], genResolve: Resolver[Repr1,Repr2], gen2: Generic.Aux[C[L2,R2], Repr2]):
+//  Resolver[C[L1,R1], C[L2,R2]] = new Resolver[C[L1,R1], C[L2,R2]] {
+//    def apply(v1: C[L1,R1]) = gen2.from(genResolve(gen1.to(v1)))
+//  }
+}
+
+trait Priority1ResolutionImplicits extends Priority01ResolutionImplicits {
   /**Resolver base cases - primitive expressions and typed variables don't need to resolve to anything.
     * Note - there is no resolver for untyped variables - they are 'resolved' by being bound.*/
   implicit def PrimitiveExprResolver[E <: PrimitiveExpr[_]]: Resolver[E,E] = Resolver.nonResolver[E]
@@ -48,111 +106,6 @@ trait StandardPriorityResolutionImplicits extends LowPriorityResolutionImplicits
   implicit def VariableResolver[V <: UntypedVariable[V],T]: Resolver[TypedVariable[T],TypedVariable[T]] =
     Resolver.nonResolver[TypedVariable[T]]
 
-  /**Standard inductive cases*/
-
-  //Ring expresssions
-  implicit def AddResolver[L <: RingExpr, R <: RingExpr, L1 <: RingExpr, R1 <: RingExpr]
-  (implicit resolve1: Resolver[L,L1], resolve2: Resolver[R,R1]): Resolver[AddExpr[L,R],AddExpr[L1,R1]] =
-    new Resolver[AddExpr[L,R],AddExpr[L1,R1]] {
-      def apply(v1: AddExpr[L,R]) = AddExpr(resolve1(v1.c1),resolve2(v1.c2))
-    }
-
-  implicit def MultiplyResolver[L <: RingExpr, R <: RingExpr, L1 <: RingExpr, R1 <: RingExpr]
-  (implicit resolve1: Resolver[L,L1], resolve2: Resolver[R,R1]): Resolver[MultiplyExpr[L,R],MultiplyExpr[L1,R1]] =
-    new Resolver[MultiplyExpr[L,R],MultiplyExpr[L1,R1]] {
-      def apply(v1: MultiplyExpr[L,R]) = MultiplyExpr(resolve1(v1.c1),resolve2(v1.c2))
-    }
-
-  implicit def DotResolver[L <: RingExpr, R <: RingExpr, L1 <: RingExpr, R1 <: RingExpr]
-  (implicit resolve1: Resolver[L,L1], resolve2: Resolver[R,R1]): Resolver[DotExpr[L,R],DotExpr[L1,R1]] =
-    new Resolver[DotExpr[L,R],DotExpr[L1,R1]] {
-      def apply(v1: DotExpr[L,R]) = DotExpr(resolve1(v1.c1),resolve2(v1.c2))
-    }
-
-  implicit def NotResolver[R1 <: RingExpr,R2 <: RingExpr](implicit resolver: Resolver[R1,R2]):
-  Resolver[NotExpr[R1],NotExpr[R2]] = new Resolver[NotExpr[R1],NotExpr[R2]] {
-    def apply(v1: NotExpr[R1]): NotExpr[R2] = NotExpr(resolver(v1.c1))
-  }
-
-  implicit def NegateResolver[R1 <: RingExpr,R2 <: RingExpr](implicit resolver: Resolver[R1,R2]):
-  Resolver[NegateExpr[R1],NegateExpr[R2]] = new Resolver[NegateExpr[R1],NegateExpr[R2]] {
-    def apply(v1: NegateExpr[R1]): NegateExpr[R2] = NegateExpr(resolver(v1.c1))
-  }
-
-  implicit def SumResolver[R1 <: RingExpr,R2 <: RingExpr](implicit resolver: Resolver[R1,R2]):
-  Resolver[SumExpr[R1],SumExpr[R2]] = new Resolver[SumExpr[R1],SumExpr[R2]] {
-    def apply(v1: SumExpr[R1]): SumExpr[R2] = SumExpr(resolver(v1.c1))
-  }
-
-  implicit def SngResolver[K <: KeyExpr,R <: RingExpr,K1 <: KeyExpr,R1 <: RingExpr]
-  (implicit resolveK: Resolver[K,K1], resolveR: Resolver[R,R1]): Resolver[SngExpr[K,R],SngExpr[K1,R1]] =
-    new Resolver[SngExpr[K,R],SngExpr[K1,R1]] { def apply(v1: SngExpr[K,R]) = SngExpr(resolveK(v1.key),resolveR(v1.value)) }
-
-  implicit def EqualsPredicateResolver[K1 <: KeyExpr,K2 <: KeyExpr,K1B <: KeyExpr,K2B <: KeyExpr]
-  (implicit resolve1: Resolver[K1,K1B], resolve2: Resolver[K2,K2B]): Resolver[EqualsPredicate[K1,K2],EqualsPredicate[K1B,K2B]] =
-    new Resolver[EqualsPredicate[K1,K2],EqualsPredicate[K1B,K2B]] {
-      def apply(v1: EqualsPredicate[K1,K2]) = EqualsPredicate(resolve1(v1.c1),resolve2(v1.c2))
-    }
-
-  implicit def IntPredicateResolver[K1 <: KeyExpr,K2 <: KeyExpr,K1B <: KeyExpr,K2B <: KeyExpr]
-  (implicit resolve1: Resolver[K1,K1B], resolve2: Resolver[K2,K2B]): Resolver[IntPredicate[K1,K2],IntPredicate[K1B,K2B]] =
-    new Resolver[IntPredicate[K1,K2],IntPredicate[K1B,K2B]] {
-      def apply(v1: IntPredicate[K1,K2]) = IntPredicate(resolve1(v1.c1),resolve2(v1.c2), v1.p, v1.opString)
-    }
-
-  implicit def ToRingResolver[R <: Expr,R1 <: Expr]
-  (implicit resolve: Resolver[R,R1]):Resolver[ToRingExpr[R],ToRingExpr[R1]] =
-    new Resolver[ToRingExpr[R],ToRingExpr[R1]] { def apply(v1: ToRingExpr[R]) = ToRingExpr(resolve(v1.c1)) }
-
-  //Key expressions
-  implicit def BoxedRingResolver[R <: Expr,R1 <: RingExpr]
-  (implicit resolve: Resolver[R,R1]):Resolver[BoxedRingExpr[R],BoxedRingExpr[R1]] =
-    new Resolver[BoxedRingExpr[R],BoxedRingExpr[R1]] { def apply(v1: BoxedRingExpr[R]) = BoxedRingExpr(resolve(v1.c1)) }
-
-  implicit def LabelResolver[R <: RingExpr,R1 <: RingExpr]
-  (implicit resolve: Resolver[R,R1]):Resolver[LabelExpr[R],LabelExpr[R1]] =
-    new Resolver[LabelExpr[R],LabelExpr[R1]] { def apply(v1: LabelExpr[R]) = LabelExpr(resolve(v1.c1)) }
-
-
-  //Tupling and projection
-  implicit def Tuple2RingResolver[K1 <: RingExpr,K2 <: RingExpr,K1B <: RingExpr,K2B <: RingExpr]
-  (implicit resolve1: Resolver[K1,K1B], resolve2: Resolver[K2,K2B]): Resolver[Tuple2RingExpr[K1,K2],Tuple2RingExpr[K1B,K2B]] =
-    new Resolver[Tuple2RingExpr[K1,K2],Tuple2RingExpr[K1B,K2B]] {
-      def apply(v1: Tuple2RingExpr[K1,K2]) = Tuple2RingExpr(resolve1(v1.c1),resolve2(v1.c2))
-    }
-
-  implicit def Tuple2KeyResolver[K1 <: KeyExpr,K2 <: KeyExpr,K1B <: KeyExpr,K2B <: KeyExpr]
-  (implicit resolve1: Resolver[K1,K1B], resolve2: Resolver[K2,K2B]): Resolver[Tuple2KeyExpr[K1,K2],Tuple2KeyExpr[K1B,K2B]] =
-    new Resolver[Tuple2KeyExpr[K1,K2],Tuple2KeyExpr[K1B,K2B]] {
-      def apply(v1: Tuple2KeyExpr[K1,K2]) = Tuple2KeyExpr(resolve1(v1.c1),resolve2(v1.c2))
-    }
-
-  implicit def Tuple2VariableResolver[K1 <: VariableExpr[K1],K2 <: VariableExpr[K2],K1B <: VariableExpr[K1B],K2B <: VariableExpr[K2B]]
-  (implicit resolve1: Resolver[K1,K1B], resolve2: Resolver[K2,K2B]): Resolver[Tuple2VariableExpr[K1,K2],Tuple2VariableExpr[K1B,K2B]] =
-    new Resolver[Tuple2VariableExpr[K1,K2],Tuple2VariableExpr[K1B,K2B]] {
-      def apply(v1: Tuple2VariableExpr[K1,K2]) = Tuple2VariableExpr(resolve1(v1.c1),resolve2(v1.c2))
-    }
-
-  implicit def Tuple3RingResolver[K1 <: RingExpr,K2 <: RingExpr,K3 <: RingExpr,K1B <: RingExpr,K2B <: RingExpr,K3B <: RingExpr]
-  (implicit resolve1: Resolver[K1,K1B], resolve2: Resolver[K2,K2B], resolve3: Resolver[K3,K3B]): Resolver[Tuple3RingExpr[K1,K2,K3],Tuple3RingExpr[K1B,K2B,K3B]] =
-    new Resolver[Tuple3RingExpr[K1,K2,K3],Tuple3RingExpr[K1B,K2B,K3B]] {
-      def apply(v1: Tuple3RingExpr[K1,K2,K3]) = Tuple3RingExpr(resolve1(v1.c1),resolve2(v1.c2),resolve3(v1.c3))
-    }
-
-  implicit def Tuple3KeyResolver[K1 <: KeyExpr,K2 <: KeyExpr,K3 <: KeyExpr,K1B <: KeyExpr,K2B <: KeyExpr,K3B <: KeyExpr]
-  (implicit resolve1: Resolver[K1,K1B], resolve2: Resolver[K2,K2B], resolve3: Resolver[K3,K3B]): Resolver[Tuple3KeyExpr[K1,K2,K3],Tuple3KeyExpr[K1B,K2B,K3B]] =
-    new Resolver[Tuple3KeyExpr[K1,K2,K3],Tuple3KeyExpr[K1B,K2B,K3B]] {
-      def apply(v1: Tuple3KeyExpr[K1,K2,K3]) = Tuple3KeyExpr(resolve1(v1.c1),resolve2(v1.c2),resolve3(v1.c3))
-    }
-
-  implicit def Tuple3VariableResolver[
-    K1 <: VariableExpr[K1],K2 <: VariableExpr[K2],K3 <: VariableExpr[K3],
-    K1B <: VariableExpr[K1B],K2B <: VariableExpr[K2B],K3B <: VariableExpr[K3B]
-  ](implicit resolve1: Resolver[K1,K1B], resolve2: Resolver[K2,K2B], resolve3: Resolver[K3,K3B]):
-    Resolver[Tuple3VariableExpr[K1,K2,K3],Tuple3VariableExpr[K1B,K2B,K3B]] =
-      new Resolver[Tuple3VariableExpr[K1,K2,K3],Tuple3VariableExpr[K1B,K2B,K3B]] {
-        def apply(v1: Tuple3VariableExpr[K1,K2,K3]) = Tuple3VariableExpr(resolve1(v1.c1),resolve2(v1.c2),resolve3(v1.c3))
-    }
 //
 //  implicit def Project1RingResolver[R1 <: RingExpr with C1Expr,R2 <: RingExpr with C1Expr](implicit resolver: Resolver[R1,R2]):
 //  Resolver[Project1RingExpr[R1],Project1RingExpr[R2]] = new Resolver[Project1RingExpr[R1],Project1RingExpr[R2]] {
@@ -183,13 +136,10 @@ trait StandardPriorityResolutionImplicits extends LowPriorityResolutionImplicits
 //  Resolver[Project3KeyExpr[R1],Project3KeyExpr[R2]] = new Resolver[Project3KeyExpr[R1],Project3KeyExpr[R2]] {
 //    def apply(v1: Project3KeyExpr[R1]): Project3KeyExpr[R2] = Project3KeyExpr(resolver(v1.c1))
 //  }
-
-
-
 }
 
 
-trait HighPriorityResolutionImplicits extends StandardPriorityResolutionImplicits {
+trait Priority2ResolutionImplicits extends Priority1ResolutionImplicits {
   //Note that the below doesnt strictly need to be here - as a there is no standalone resolver for an infinite mapping,
   //the standard MultiplyResolver if tried for MultiplyExpr with an InfMapping on the RHS will fail. However, it's logical
   //to aid the compiler by making it try this one first for such expressions.
@@ -394,4 +344,4 @@ trait HighPriorityBindingImplicits extends StandardPriorityBindingImplicits {
   implicit def TypedNonBinder[V <: VariableExpr[V],V1 <: TypedVariable[_],T]: Binder[V,T,V1,V1] = Binder.nonBinder[V,T,V1]
 }
 
-trait VariableResolutionImplicits extends HighPriorityBindingImplicits with HighPriorityResolutionImplicits
+trait VariableResolutionImplicits extends HighPriorityBindingImplicits with Priority2ResolutionImplicits
