+++ src/test/scala/IntegrationTests.scala
@@ -1,44 +1,50 @@
-//package slender
+package slender
+
+import org.apache.spark.sql.SparkSession
+import org.scalatest.FunSuite
+import scala.collection.immutable.Map
+
+class IntegrationTests extends FunSuite {
+
+  import implicits._
+
+//  val spark = SparkSession.builder()
+//    .appName("test")
+//    .config("spark.master", "local")
+//    .getOrCreate()
 //
-//import org.apache.spark.sql.SparkSession
-//import org.scalatest.FunSuite
-//import scala.collection.immutable.Map
+//  implicit val sc = spark.sparkContext
+
+  val bagOfInts = PhysicalCollection(
+    Map(1 -> 1, 2 -> 2, 3 -> 3)
+  )
+
+//  test("Nested yield") {
+//    val query = For (X <-- bagOfInts) Yield toK(For(Y <-- bagOfInts) Yield Y)
+//    assert(query.isResolved)
+//    assert(query.isEvaluable)
+//  }
 //
-//class IntegrationTests extends FunSuite with TestUtils {
-//
-//  import implicits._
-//
-////  val spark = SparkSession.builder()
-////    .appName("test")
-////    .config("spark.master", "local")
-////    .getOrCreate()
-////
-////  implicit val sc = spark.sparkContext
-//
-//  val bagOfInts = PhysicalCollection(
-//    Map(1 -> 1, 2 -> 2, 3 -> 3)
-//  )
-//
-////  test("Nested yield") {
-////    val query = For (X <-- bagOfInts) Yield toK(For(Y <-- bagOfInts) Yield Y)
-////    assert(query.isResolved)
-////    assert(query.isEvaluable)
-////  }
-////
-////  test("Predicated yield test") {
-////    val query = For (X <-- bagOfInts iff X === X) Yield X
-////    assert(query.isResolved)
-////    //query.resolve
-////    println(query.eval)
-////  }
-//
+//  test("Predicated yield test") {
+//    val query = For (X <-- bagOfInts iff X === X) Yield X
+//    assert(query.isResolved)
+//    //query.resolve
+//    println(query.eval)
+//  }
+
+  test("") {
+//    val expr = GenBinaryExpr(IntKeyExpr(1), IntKeyExpr(1))
+    val expr = SumExpr(PhysicalCollection(Set(1,2,3)))
+    expr.resolve
+    println(expr.resolve.eval)
+  }
 //  test("Simple query") {
 //    val query = For(X <-- bagOfInts) Yield X --> NumericExpr(1)
 //    assert(query.isResolved)
 //    assert(query.isEvaluable)
 //    println(query.eval)
 //  }
-//
+
 //  test("Full packet query") {
 //    val serverData = Map(
 //      (1,
@@ -104,4 +110,4 @@
 //    assert(query.isResolved)
 //    assert(query.isEvaluable)
 //  }
-//}
+}
